+++ src/app.rs
@@ -12047,6 +12047,7 @@ Error: `{err}`"
                 reporter_corpus_export: None,
                 reporter_construct_handoff: None,
                 uniprot_projection_audit: None,
+                uniprot_linked_transcript_inventory: None,
                 uniprot_projection_audit_parity: None,
                 lab_assistant_instructions: None,
                 feature_location_edit_report: None,